Observe document metadata changes and notify sidebar
In order to ensure that the document metadata attached to new
annotations and stored in the sidebar matches the actual document
metadata, the client needs to observe changes to `<meta>` and `<link>`
tags in the host page.

In addition to the document metadata, the URL itself can depend on
the metadata because `<link rel="canonical">` overrides the actual URL
reported by `window.location.href`.

A small amount of buffering is used when subscribing to these changes
to avoid triggering multiple annotation search requests in the sidebar
if the web app updates the URL and metadata tags over the course of
several event loop turns.